nix-private sha: 7f1d281

+++ flake/modules/home-manager/nixos/graphical/webapps/omarchy-webapp-handler-zoom.sh
@@ -18,12 +18,16 @@ if [[ $url =~ ^zoom(mtg|us):// ]]; then
   fi
 fi
 
-hyprctl dispatch workspace 19
+compositor="$($HOME/Scripts/nixos/compositor)"
+if [[ "$compositor" == "niri" ]]; then
+  niri msg action focus-workspace zoom
+else
+  hyprctl dispatch workspace 19
 
-existing_zoom=$(hyprctl clients -j | jq -r '.[] | select(.class | test("chrome-app.zoom")) | .pid')
-
-if [[ -n "$existing_zoom" ]]; then
-  kill "$existing_zoom"
+  existing_zoom=$(hyprctl clients -j | jq -r '.[] | select(.class | test("chrome-app.zoom")) | .pid')
+  if [[ -n "$existing_zoom" ]]; then
+    kill "$existing_zoom"
+  fi
 fi
 
 exec omarchy-launch-webapp "$web_url"
